Move metrics to own file from layered_repository/timeline.rs
